src/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
idna==3.3
psycopg2==2.9.2
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
uWSGI==2.0.20
